# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PT Multi Daya Mitra Company Profile CMS API
  version: 1.2.0
  summary: REST API for the public corporate website and protected CMS.
  description: |
    Sign-in is two-step: POST /auth/login verifies the password and returns
    either tokens (2FA disabled or a valid trusted-device token was presented)
    or an OTP challenge completed via POST /auth/verify-otp. Analytics
    collection is first-party and feature-flagged via the "analytics" setting.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker API
security: []
...
paths:
...
  /public/pages/{key}:
    get:
      tags: [Public]
      summary: Get CMS-managed static page content.
      parameters:
        - $ref: "#/components/parameters/PageKey"
      responses:
        "200":
          description: Published page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /public/news:
    get:
      tags: [Public]
      summary: List published news with pagination and optional category filter.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated news.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsList"
  /public/news/{slug}:
    get:
      tags: [Public]
      summary: Get published news detail by slug.
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: News detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/careers:
    get:
      tags: [Public]
      summary: List published job openings.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated careers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerList"
  /public/careers/{slug}:
    get:
      tags: [Public]
      summary: Get published job detail by slug.
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Career detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
        "404":
          $ref: "#/components/responses/NotFound"
  /public/contacts:
    post:
      tags: [Public]
      summary: Submit contact inquiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /public/analytics/collect:
    post:
      tags: [Analytics]
      summary: Ingest a batch of tracker events (non-blocking, buffered).
      description: |
        Always answers before any database work. Bots (user-agent), admin
        sessions, and Do-Not-Track requests are dropped server-side. Content
        type may be text/plain so sendBeacon avoids a CORS preflight.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsCollectInput"
      responses:
        "202":
          description: Accepted (or silently dropped by policy).
        "204":
          description: Analytics is disabled.
        "429":
          description: Rate limited.
  /public/redirects/resolve:
    get:
      tags: [Redirects]
      summary: Resolve a branded short-link slug (hot path, memory-cached).
      description: |
        Called by the frontend middleware for company.com/{slug}. The visitor's
        forwarded headers feed asynchronous scan analytics; the response never
        waits for logging. The middleware issues the actual 301/302.
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Live target.
          content:
            application/json:
              schema:
                type: object
                properties:
                  found: { type: boolean }
                  destination: { type: string }
                  redirectType: { type: integer, enum: [301, 302] }
        "404":
          description: Unknown, inactive, or expired slug ({ found:false }).
  /auth/login:
    post:
      tags: [Auth]
      summary: "Step 1: verify email + password."
      description: |
        Returns status "ok" with tokens when 2FA is disabled or a valid
        trusted-device token (matching the stored device fingerprint) is
        presented; otherwise opens an email-code challenge and returns
        status "otp_required".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Tokens or a pending OTP challenge.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: "Step 3: exchange the emailed code for a session."
      description: |
        Codes are single-use and expire per the security settings (default 5
        minutes). Wrong attempts are counted per challenge and lock it once
        the budget is spent. With trustDevice=true the response includes an
        opaque trusted-device token to store in an HTTP-only cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPInput"
      responses:
        "200":
          description: Session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOTPResult"
        "401":
          description: Invalid, expired, or already-used code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/resend-otp:
    post:
      tags: [Auth]
      summary: Re-send a fresh code for a pending challenge.
      description: Enforces a cooldown and a per-challenge resend cap; a new code replaces the old one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId]
              properties:
                challengeId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: New code sent; challenge details refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset code by email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Always accepted (never reveals whether the account exists).
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password using the emailed reset code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeInput"
      responses:
        "200":
          description: Password updated.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/verify-invite:
    post:
      tags: [Auth]
      summary: Activate an invited account with the emailed code and a new password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeInput"
      responses:
        "200":
          description: Account activated; tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue a new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "204":
          description: Logged out.
  /admin/dashboard:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: CMS dashboard counts.
      responses:
        "200":
          description: Dashboard metrics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  counts:
                    type: object
                    additionalProperties:
                      type: integer
  /admin/navigation:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Get the editable menu tree with its version.
      responses:
        "200":
          description: Menu items + version for optimistic locking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationMenu"
    put:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Replace the menu tree (two levels max, validated + normalized).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavigationMenu"
      responses:
        "200":
          description: Saved menu.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationMenu"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/settings:
    get:
      tags: [Settings]
      security:
        - bearerAuth: []
      summary: List all settings values keyed by name.
      responses:
        "200":
          description: Map of setting key to raw JSON value.
  /admin/settings/{key}:
    get:
      tags: [Settings]
      security:
        - bearerAuth: []
      summary: Get one setting with its version (missing keys return {} at version 0).
      parameters:
        - $ref: "#/components/parameters/SettingKey"
      responses:
        "200":
          description: Setting document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
    put:
      tags: [Settings]
      security:
        - bearerAuth: []
      summary: Upsert an editable setting under optimistic locking.
      description: "Editable keys: site, analytics, security. The navigation key has its own endpoint and is rejected here."
      parameters:
        - $ref: "#/components/parameters/SettingKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, version]
              properties:
                value:
                  type: object
                  additionalProperties: true
                version:
                  type: integer
                  description: 0 to create; current version to update.
      responses:
        "200":
          description: Saved setting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/media:
    get:
      tags: [Media]
      security:
        - bearerAuth: []
      summary: List uploaded media files, newest first.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          schema:
            type: string
          description: Filter by file name.
      responses:
        "200":
          description: Paginated media.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaList"
  /admin/media/upload:
    post:
      tags: [Media]
      security:
        - bearerAuth: []
      summary: Upload a file (multipart form, field "file", max 25MB).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Stored file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
  /admin/media/{id}:
    delete:
      tags: [Media]
      security:
        - bearerAuth: []
      summary: Permanently delete a media file (row + stored object).
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/profile/devices:
    get:
      tags: [Profile]
      security:
        - bearerAuth: []
      summary: List the signed-in user's trusted devices.
      responses:
        "200":
          description: Active trusted devices.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrustedDevice"
  /admin/profile/devices/{id}:
    delete:
      tags: [Profile]
      security:
        - bearerAuth: []
      summary: Revoke one of the user's own trusted devices.
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "204":
          description: Revoked.
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/profile/devices/revoke-all:
    post:
      tags: [Profile]
      security:
        - bearerAuth: []
      summary: Revoke all of the user's trusted devices.
      responses:
        "204":
          description: Revoked.
  /admin/profile/login-history:
    get:
      tags: [Profile]
      security:
        - bearerAuth: []
      summary: Recent sign-in events for the current user, including failures.
      responses:
        "200":
          description: Audit-log entries (entity_type=auth).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/LoginHistoryEntry"
  /admin/analytics/dashboard:
    get:
      tags: [Analytics]
      security:
        - bearerAuth: []
      summary: Full analytics dashboard payload for a date range (cached ~60s).
      parameters:
        - $ref: "#/components/parameters/AnalyticsRange"
        - name: interval
          in: query
          schema:
            type: string
            enum: [hour, day, week, month, year]
            default: day
        - name: device
          in: query
          schema: { type: string, enum: [desktop, mobile, tablet] }
        - name: country
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string, enum: [direct, organic, referral, social] }
      responses:
        "200":
          description: Overview, timeseries, breakdowns, pages, events, vitals, API stats.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsDashboard"
  /admin/analytics/realtime:
    get:
      tags: [Analytics]
      security:
        - bearerAuth: []
      summary: Live picture (last 5 minutes) from in-memory state.
      responses:
        "200":
          description: Active visitors, current pages, recent events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsRealtime"
  /admin/analytics/admin-activity:
    get:
      tags: [Analytics]
      security:
        - bearerAuth: []
      summary: Admin usage counters (logins, failed logins, content changes) + recent audit rows.
      parameters:
        - $ref: "#/components/parameters/AnalyticsRange"
      responses:
        "200":
          description: Admin activity summary.
  /admin/analytics/options:
    get:
      tags: [Analytics]
      security:
        - bearerAuth: []
      summary: Filter dropdown options (recent paths and countries).
      responses:
        "200":
          description: Options.
  /admin/analytics/export:
    get:
      tags: [Analytics]
      security:
        - bearerAuth: []
      summary: Export the dashboard as CSV (one report) or XLSX (full workbook).
      parameters:
        - $ref: "#/components/parameters/AnalyticsRange"
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, xlsx]
            default: csv
        - name: report
          in: query
          description: CSV only.
          schema:
            type: string
            enum: [summary, timeseries, pages, events, breakdowns]
            default: summary
      responses:
        "200":
          description: File download (text/csv or application/vnd.openxmlformats-officedocument.spreadsheetml.sheet).
  /admin/redirects:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: List short links (q search; status active|inactive|expired).
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, expired] }
      responses:
        "200":
          description: Paginated redirects with lifetime scan totals.
    post:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Create a short link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedirectInput"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redirect"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/redirects/dashboard:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Top scanned links, scan trends, recent scans, geo distribution.
      parameters:
        - $ref: "#/components/parameters/AnalyticsRange"
      responses:
        "200":
          description: Dashboard payload.
  /admin/redirects/export:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Export every live short link (with scan totals) as CSV.
      responses:
        "200":
          description: CSV download.
  /admin/redirects/bulk-delete:
    post:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Archive up to 200 short links at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: "{ archived: n }"
  /admin/redirects/{id}:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Get one short link.
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "200":
          description: Redirect.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redirect"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Update a short link (optimistic locking).
      parameters:
        - $ref: "#/components/parameters/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedirectInput"
      responses:
        "200":
          description: Updated.
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Archive a short link (stops redirecting; scan history kept).
      parameters:
        - $ref: "#/components/parameters/UUID"
        - name: version
          in: query
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Archived.
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/redirects/{id}/duplicate:
    post:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Copy a short link as an inactive draft with a derived slug.
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "201":
          description: The copy.
  /admin/redirects/{id}/stats:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Scan analytics for one link (totals, uniques, trend, breakdowns, history).
      parameters:
        - $ref: "#/components/parameters/UUID"
        - $ref: "#/components/parameters/AnalyticsRange"
      responses:
        "200":
          description: Redirect + stats + short URL.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageKey:
      name: key
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
        description: "System pages: home, about, contact, services, products, news, career — plus any admin-created page slug."
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  schemas:
...
    Page:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        title:
          type: string
        content:
          type: object
          additionalProperties: true
        status:
          type: string
        publishedAt:
          type: [string, "null"]
          format: date-time
        seo:
          $ref: "#/components/schemas/SEO"
        version:
          type: integer
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...